feat: refactor to have separate dataset utils and partition processor
